gestion des menus et creation des dashboards

+++ database/migrations/2022_05_27_091113_create_notes_table.php
@@ -17,7 +17,7 @@ class CreateNotesTable extends Migration
             $table->id();
             $table->foreignId('etudiant_id')->constrained()->onDelete('cascade');
             $table->foreignId('evaluation_id')->constrained()->onDelete('cascade');
-            $table->double('note');
+            $table->float('note');
             $table->timestamps();
         });
     }
